requirements.txt
# pysounds requirements


# development utilities and general tools
black>=24.8
Pint>=0.24
isort==5.0.2
mypy>=1.13
pytest==4.5.0
toml

# Specific requirements
nidaqmx
matplotlib>=3.9
pandas
seaborn
nose==1.3.7
numpy<1.2
psutil>=6.1.0
pyparsing==2.4.0
resampy>=0.4
scipy>1.5.0
pyaudio  # this will have to be done manually
# get pyaudio win-amd64.whl file from https://www.lfd.uci.edu/~gohlke/pythonlibs/
# make a downloads file in pysounds, and put it in the downloads directory
# 11/2024: Gohlke's site is permanently down. 
# 
pyqt6>=6.7
pyqt6-sip>=13.8
pyqtgraph>=0.13.7
tdtpy
# pywin32

# -e git://github.com/pbmanis/pylibrary.git#egg=pylibrary
# -e C:/Users/experimenters/Desktop/Python/pylibrary
-e /Users/pbmanis/Desktop/Python/pylibrary